Added system-based "Hello World" example
